requirements.txt
Django==1.6.8
httplib2==0.7.7
restclient==0.11.0
feedparser==4.1
Markdown==2.3.1
simplejson==3.4.0
smartypants==1.6.0.2
uuid==1.30
psycopg2==2.5.1
nose==1.3.4
versiontools==1.9.1
statsd==2.0.2
pep8==1.4
pyflakes==0.5.0
flake8==1.7.0
lxml==3.2.3
fuzzywuzzy==0.1
sure==1.1.7
ipython==2.3.0
ipdb==0.8
extras==0.0.3
python-mimeparse==0.1.4
testtools==0.9.35
testscenarios==0.4
python-subunit==0.0.18
lettuce==0.2.19ccnmtl
rdflib==3.1.0
selenium==2.35.0
coverage==3.5.3
requests==0.13.3
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==0.26.0
six==1.2.0
cssselect==0.8
BeautifulSoup==3.2.1
Pillow==2.3.0
factory_boy==2.4.1
sqlparse==0.1.11

djangowind==0.13.2
sorl==3.1
tagging==0.3-pre
typogrify==1.0
django-autoslug==1.7.2
django-indexer==0.2.1
django-templatetag-sugar==0.1
django-paging==0.2.2
South==1.0
django-annoying==0.7.6
django-nose==1.2
django-appconf==0.4.1
django_compressor==1.3
django-statsd-mozilla==0.3.11
raven==4.0.3
django-bootstrap-form==2.0.3
django-debug-toolbar==1.1
django-waffle==0.9.0
django-jenkins==0.16.3
django-smoketest==0.2.2
django-extensions==1.1.0
django-stagingcontext==0.1.0
django-impersonate==0.8.0
django-treebeard==2.0b2
django-markwhat==1.0
django-pagetree==1.0.1
django-pageblocks==0.5.11
django-quizblock==0.9.4
django-pagetree_export==0.4
djangorestframework==2.4.3
numpy==1.7.1
